[W-18338037] refactor: metadataDescribe without cli (#6401)
* feat: first attempt at switching describeMetadata.ts to not use CLI - works when the reload button in Org Browser is clicked on a single org

* fix: add a listener when the org changes

* refactor: use createDirectory() and writeFile() instead of node:fs

* Merge branch 'develop' into daphne/W-18338037-metadataDescribe-without-cli

* Merge branch 'develop' into daphne/W-18338037-metadataDescribe-without-cli

* Merge branch 'develop' into daphne/W-18338037-metadataDescribe-without-cli

* refactor: reduce mutation

* refactor: remove from barrel

* refactor: describeMetadata() returns Promise<DescribeMetadataResult> instead of Promise<string>

* Merge branch 'develop' into daphne/W-18338037-metadataDescribe-without-cli

* refactor: handling null suffix and reducing code path complexity
