+++ packages/salesforcedx-vscode-core/src/orgBrowser/metadataType.ts
@@ -14,7 +14,7 @@ import {
   fileOrFolderExists
 } from '@salesforce/salesforcedx-utils-vscode';
 import * as path from 'node:path';
-import { describeMetadata } from '../commands';
+import { describeMetadata, DescribeMetadataResult } from '../commands/describeMetadata';
 import { coerceMessageKey, nls } from '../messages';
 import { telemetryService } from '../telemetry';
 
@@ -26,9 +26,9 @@ export type MetadataObject = {
   xmlName: string;
   label: string;
 };
+
 export class TypeUtils {
   public static readonly FOLDER_TYPES = new Set(['CustomObject', 'Dashboard', 'Document', 'EmailTemplate', 'Report']);
-
   public static readonly UNSUPPORTED_TYPES = new Set(['InstalledPackage', 'Profile', 'Scontrol']);
 
   public getTypesFolder(): string {
@@ -37,22 +37,19 @@ export class TypeUtils {
       telemetryService.sendException('metadata_type_workspace', err);
       throw new Error(err);
     }
-    const metadataTypesPath = projectPaths.metadataFolder();
-    return metadataTypesPath;
+    return projectPaths.metadataFolder();
   }
 
   public async loadTypes(forceRefresh?: boolean): Promise<MetadataObject[]> {
     const typesFolder = this.getTypesFolder();
     const typesPath = path.join(typesFolder, 'metadataTypes.json');
 
-    let typesList: MetadataObject[];
-    if (forceRefresh || !(await fileOrFolderExists(typesPath))) {
-      const result = await describeMetadata(typesFolder);
-      typesList = await buildTypesList({ metadataJSONContents: result });
-    } else {
-      typesList = await buildTypesList({ metadataTypesPath: typesPath });
-    }
-    return typesList;
+    const describeResult =
+      forceRefresh || !(await fileOrFolderExists(typesPath))
+        ? await describeMetadata(typesFolder)
+        : // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
+          (JSON.parse(await readFile(typesPath)) as DescribeMetadataResult);
+    return buildTypesList(describeResult);
   }
 
   public getFolderForType(metadataType: string): string {
@@ -67,21 +64,16 @@ export class TypeUtils {
   }
 }
 
-const buildTypesList = async (
-  input: { metadataTypesPath: string } | { metadataJSONContents: string }
-): Promise<MetadataObject[]> => {
+const buildTypesList = (describeResult: DescribeMetadataResult): MetadataObject[] => {
   try {
-    const jsonObject = JSON.parse(
-      'metadataJSONContents' in input ? input.metadataJSONContents : await readFile(input.metadataTypesPath)
-    );
-    // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
-    const metadataTypeObjects = (jsonObject.result.metadataObjects as MetadataObject[])
+    const metadataTypeObjects = describeResult.metadataObjects
       .filter(type => !isNullOrUndefined(type.xmlName) && !TypeUtils.UNSUPPORTED_TYPES.has(type.xmlName))
       .map(mdTypeObject => ({
         ...mdTypeObject,
         label: nls.localize(coerceMessageKey(mdTypeObject.xmlName)).startsWith(MISSING_LABEL_MSG)
           ? mdTypeObject.xmlName
-          : nls.localize(coerceMessageKey(mdTypeObject.xmlName))
+          : nls.localize(coerceMessageKey(mdTypeObject.xmlName)),
+        suffix: mdTypeObject.suffix ?? undefined
       }))
       .sort((a, b) => (a.label > b.label ? 1 : -1));
     telemetryService.sendEventData('Metadata Types Quantity', undefined, {
